+++ src/structures/binaries/delete.c
@@ -6,7 +6,7 @@
 /*   By: dlaurent <dlaurent@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/30 18:01:59 by dlaurent          #+#    #+#             */
-/*   Updated: 2018/10/03 20:14:30 by dlaurent         ###   ########.fr       */
+/*   Updated: 2018/10/17 14:52:02 by dlaurent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ void		bin_delete_specified_item(t_bin_item *item)
 void		bin_delete_item(t_bin *bin, const char *key)
 {
 	int			i;
-	int			index;
+	long		index;
 	t_bin_item	*item;
 
 	i = 1;
